Add search input to admin user list
